+++ src/index.ts
@@ -1,3 +1,4 @@
+import "./plugins/autoRegister.js";
 export { collectPullRequests } from "./collectors/pullRequests.js";
 export { calculateCycleTime } from "./calculators/cycleTime.js";
 export { calculateReviewMetrics } from "./calculators/reviewMetrics.js";